+++ README.md
@@ -6,7 +6,7 @@
 ## 2단계 기능 요구사항
 - [ ] 리팩터링 힌트
   - [x] 메서드 추출
-  - [ ] UpdateJdbcTemplate, InsertJdbcTemplate 클래스 추출
+  - [x] UpdateJdbcTemplate, InsertJdbcTemplate 클래스 추출
   - [ ] 템플릿 메서드 패턴 적용
   - [ ] 도메인 의존도 제거
   - [ ] SelectJdbcTemplate 클래스 추출
