+++ app/src/main/java/com/techcourse/dao/UpdateJdbcTemplate.java
@@ -1,0 +1,57 @@
+package com.techcourse.dao;
+
+import com.techcourse.domain.User;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import javax.sql.DataSource;
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.SQLException;
+
+public class UpdateJdbcTemplate {
+
+    private static final Logger log = LoggerFactory.getLogger(UpdateJdbcTemplate.class);
+
+    private String createQueryForUpdate() {
+        return "update users set account=?, password=?, email=? where id=?";
+    }
+
+    private void setValueForUpdate(User user, PreparedStatement pstmt) throws SQLException {
+        pstmt.setString(1, user.getAccount());
+        pstmt.setString(2, user.getPassword());
+        pstmt.setString(3, user.getEmail());
+        pstmt.setLong(4, user.getId());
+        pstmt.executeUpdate();
+    }
+
+    public void update(User user, UserDao userDao) {
+        Connection conn = null;
+        PreparedStatement pstmt = null;
+        DataSource dataSource = userDao.getDataSource();
+        try {
+            String sql = createQueryForUpdate();
+            conn = dataSource.getConnection();
+            pstmt = conn.prepareStatement(sql);
+
+            log.debug("query : {}", sql);
+
+            setValueForUpdate(user, pstmt);
+        } catch (SQLException e) {
+            log.error(e.getMessage(), e);
+            throw new RuntimeException(e);
+        } finally {
+            try {
+                if (pstmt != null) {
+                    pstmt.close();
+                }
+            } catch (SQLException ignored) {}
+
+            try {
+                if (conn != null) {
+                    conn.close();
+                }
+            } catch (SQLException ignored) {}
+        }
+    }
+}
